+++ frontend/public/js/editorSelect.js
@@ -7,7 +7,7 @@ import {
   getMode,
 } from "./contextMenu.js";
 import { contextMenu, editor } from "./components.js";
-import { selectImageButton } from "./imageSelect.js";
+import { reselectImageButton, selectImageButton } from "./imageSelect.js";
 import { uploadImageButton } from "./imageUpload.js";
 import { deleteLink } from "./components.js";
 import { saveFile } from "./fileSave.js";
@@ -22,8 +22,6 @@ document.addEventListener("selectionchange", (e) => {
     const selectedText = selection.toString();
     if (!selectedText || /^\s+$/.test(selectedText)) return;
     const range = selection.getRangeAt(0);
-    const selectedNode = range.commonAncestorContainer;
-    if ($(editor).find(selectedNode).length === 0) return;
     const rect = range.getBoundingClientRect();
     clearMenu();
     appendItem(selectImageButton());
@@ -38,20 +36,66 @@ $(editor).on("contextmenu", "a.view", (e) => {
   const anchor = e.target;
   window.getSelection().selectAllChildren(anchor);
   clearMenu();
-  appendItem(
-    $("<button>", {
-      text: "Delete Link",
-      class: "dropdown-item",
-      on: {
-        click: () => {
-          deleteLink(anchor);
-          saveFile();
-          $(contextMenu).trigger("complete");
-        },
-      },
-    })
-  );
+  appendItem(deleteLinkButton(anchor));
+  appendItem(reselectImageButton(anchor));
   setMode(MODE.MENU);
   const rect = anchor.getBoundingClientRect();
   showMenuAtPos(rect);
 });
+
+export const validateRange = (range) => {
+  if (!range || !(range instanceof Range)) {
+    console.error("Invalid range");
+    return false;
+  }
+  let valid = true;
+  $(editor)
+    .find("a.view")
+    .each((i, anchor) => {
+      if (range.intersectsNode(anchor)) {
+        valid = false;
+      }
+    });
+  return valid;
+};
+
+export const textToLink = (href, className) => {
+  const selection = window.getSelection();
+  if (!selection) {
+    console.error("No selection found");
+    return false;
+  }
+  if (selection.rangeCount === 0) {
+    console.error("No range selected");
+    return false;
+  }
+  const selectedText = selection.toString();
+  const range = selection.getRangeAt(0);
+  if (!validateRange(range)) {
+    alert(
+      "Selection must be text only. Right click to turn links back into text."
+    );
+    return false;
+  }
+  const $anchor = $("<a>", {
+    href: href,
+    html: selectedText,
+    class: className,
+  });
+  range.deleteContents();
+  range.insertNode($anchor[0]);
+  return true;
+};
+
+const deleteLinkButton = (anchor) =>
+  $("<button>", {
+    text: "Delete Link",
+    class: "dropdown-item",
+    on: {
+      click: () => {
+        deleteLink(anchor);
+        saveFile();
+        $(contextMenu).trigger("complete");
+      },
+    },
+  });