+++ frontend/public/js/imageSelect.js
@@ -1,9 +1,9 @@
-import { editor, fileSelector } from "./components.js";
+import { contextMenu, editor, fileSelector } from "./components.js";
 import { appendItem, clearMenu, showMenu } from "./contextMenu.js";
-import { deleteLink } from "./components.js";
+import { textToLink } from "./editorSelect.js";
 import { saveFile } from "./fileSave.js";
 
-const openSelectWindow = async () => {
+const openSelectWindow = async (clickHandle) => {
   const fileID = $(fileSelector).val();
   if (!fileID) {
     console.error("Invalid file ID");
@@ -16,7 +16,7 @@ const openSelectWindow = async () => {
       width: "50vw",
     },
     html: html,
-  }).on("click", handleSelectImage);
+  }).on("click", clickHandle);
   clearMenu();
   appendItem($selectionWindow);
   showMenu();
@@ -31,35 +31,45 @@ const handleSelectImage = (e) => {
     console.error("Invalid file or image ID");
     return;
   }
-  const selection = window.getSelection();
-  if (!selection) {
-    console.error("No selection found");
+  if (!textToLink(`/access/${fileID}/${imageID}`, "view")) {
     return;
   }
-  if (selection.rangeCount === 0) {
-    console.error("No range selected");
-    return;
-  }
-  const selectedText = selection.toString();
-  const range = selection.getRangeAt(0);
-  $(editor)
-    .find("a.view")
-    .each((i, anchor) => {
-      if (range.intersectsNode(anchor)) deleteLink(anchor);
-    });
-  const $anchor = $("<a>", {
-    href: `/access/${fileID}/${imageID}`,
-    html: selectedText,
-    class: "view",
-  });
-  range.deleteContents();
-  range.insertNode($anchor[0]);
   saveFile();
+  $(contextMenu).trigger("complete");
+};
+
+const handleReselectImage = (anchor, e) => {
+  if (!anchor || !anchor.href) {
+    console.error("Invalid anchor");
+    return;
+  }
+  const target = e.target.closest(".id-selectable");
+  if (!target) return;
+  const fileID = $(fileSelector).val();
+  const imageID = target.id;
+  if (!fileID || !imageID) {
+    console.error("Invalid file or image ID");
+    return;
+  }
+  anchor.href = `/access/${fileID}/${imageID}`;
+  saveFile();
+  $(contextMenu).trigger("complete");
 };
 
 export const selectImageButton = () => {
   return $("<button>", {
     text: "Select Image",
     class: "dropdown-item select-image",
-  }).on("click", openSelectWindow);
+  }).on("click", () => openSelectWindow(handleSelectImage));
+};
+
+export const reselectImageButton = (anchor) => {
+  return $("<button>", {
+    text: "Reselect Image",
+    class: "dropdown-item reselect-image",
+  }).on("click", () => {
+    openSelectWindow((e) => {
+      handleReselectImage(anchor, e);
+    });
+  });
 };