+++ frontend/public/js/imageUpload.js
@@ -1,11 +1,15 @@
 import {
   contextMenu,
+  deleteLink,
+  editor,
   fileSelector,
   imageInput,
   imagePanel,
   input,
   uploadImage,
 } from "./components.js";
+import { textToLink, validateRange } from "./editorSelect.js";
+import { saveFile } from "./fileSave.js";
 
 if (!$(contextMenu).length) {
   console.error("Context menu not found");
@@ -20,62 +24,48 @@ if (!$(input).length) {
 }
 
 $(document).on("input", imageInput, async (e) => {
-  const fileID = $(fileSelector).val();
-  if (!fileID) {
-    console.error("Invalid file ID");
-    return;
-  }
-  const file = e.target.files[0];
-  if (!file) {
-    console.error("No file selected");
-    return;
-  }
-  const formData = new FormData();
-  formData.append("image", file);
-  const response = await fetch(`/upload/${fileID}`, {
-    method: "POST",
-    body: formData,
-  });
-  if (response.ok) {
-    const imageID = await response.text();
-    $("a.view.loading")
-      .attr("href", `/access/${fileID}/${imageID}`)
-      .removeClass("loading");
-    saveFile();
-  } else {
-    console.error(response.statusText);
-  }
+  try {
+    const fileID = $(fileSelector).val();
+    if (!fileID) {
+      throw new Error("Invalid file ID");
+    }
+    const file = e.target.files[0];
+    if (!file) {
+      throw new Error("No file selected");
+    }
+    const formData = new FormData();
+    formData.append("image", file);
+    const response = await fetch(`/upload/${fileID}`, {
+      method: "POST",
+      body: formData,
+    });
+    if (response.ok) {
+      const imageID = await response.text();
+      $("a.view.loading")
+        .attr("href", `/access/${fileID}/${imageID}`)
+        .removeClass("loading");
+      saveFile();
+    } else {
+      throw new Error(response.statusText);
+    }
 
-  $(this).val("");
+    $(this).val("");
 
-  const response2 = await fetch(`/access/${fileID}`);
-  if (response2.ok) {
-    const html = await response2.text();
-    console.log(html);
-    $(imagePanel).html(html);
+    const response2 = await fetch(`/access/${fileID}`);
+    if (response2.ok) {
+      const html = await response2.text();
+      $(imagePanel).html(html);
+    }
+  } catch (error) {
+    console.error(error);
+    $(this).val("");
   }
 });
 
 const handleUploadImage = () => {
-  const selection = window.getSelection();
-  const selectedText = selection?.toString();
-  if (
-    selection &&
-    selection.rangeCount > 0 &&
-    selectedText &&
-    !/^\s+$/.test(selectedText)
-  ) {
-    const range = selection.getRangeAt(0);
-    const $anchor = $("<a>", {
-      href: `/res/image.svg`,
-      html: selectedText,
-      class: "view loading",
-    });
-
-    range.deleteContents();
-    range.insertNode($anchor[0]);
+  if (!textToLink("/res/image.svg", "view loading")) {
+    return;
   }
-
   $(input)
     .attr({
       name: "image",
@@ -85,9 +75,7 @@ const handleUploadImage = () => {
   $(contextMenu).trigger("complete");
 };
 
-$(uploadImage).on("click", async () => {
-  handleUploadImage();
-});
+$(uploadImage).on("click", handleUploadImage);
 
 export const uploadImageButton = () => {
   return $("<button>", {